Split out components into individual files + various improvements
* Add reprs
* Pass through env_markers as a dict
* Add context manager to RequirementsFile class
